correct gameLogic

+++ core/src/Entities/Creatures/Position.java
@@ -1,27 +1,25 @@
 package Entities.Creatures;
 
 public class Position {
-    float x;
-    float y;
+    int x;
+    int y;
 
-    public Position(float x, float y) {
+    public Position(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public void setX(float x) {
+    public void setX(int x) {
         this.x = x;
     }
-
-    public void setY(float y) {
+    public void setY(int y) {
         this.y = y;
     }
 
-    public float getX() {
+    public int getX() {
         return x;
     }
-
-    public float getY() {
+    public int getY() {
         return y;
     }
 }
